update utility + copy
